pyproject.toml
[tool.poetry]
name = "dj-rcp"
version = "0.1.0"
description = ""
authors = ["Nat Gordon <nat@nattyg93.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^8.1.2"
PyYAML = "^5.4.1"
celery = { version = "^4.3", extras = ["redis"] }
celery-prometheus-exporter = "^1.7"
dj-rest-auth = "^2.1.3"
django = "^2.2"
django-allauth = "^0.39.1"
django-anymail = { version = "^6.1", extras = ["mailgun"] }
django-axes = "^5.0"
django-celery-beat = "^1.5"
django-celery-results = "^1.1"
django-cors-headers = "^3.1"
django-environ = "^0.4.5"
django-extensions = "^2.2"
django-filter = "^2.2"
django-redis = "~=4.10.0" # does not follow semver
django-storages = { version = "^1.7", extras = ["boto3"] }
djangorestframework = "~=3.12.2" # does not follow semver
djangorestframework-filters = ">=1.0.0.dev0"
djangorestframework-jsonapi = "^4.0.0"
gunicorn = "^19.9"
psycopg2 = "^2.8"
python-dateutil = "^2.8"
pytz = "^2021.1"
sentry-sdk = "^1.1.0"
uritemplate = "^3.0.1"

[tool.poetry.dev-dependencies]
# debugging
django-debug-toolbar = "^2.0"
ipdb = "^0.11.0"
ipython = "^7.2"
ptpython = "^2.0"
# linting
autopep8 = "^1.4"
django-stubs = "~=1.8.0"
mypy = "^0.812"
pycodestyle = "^2.4"
pydocstyle = "^5.0.2"
pylint = "^2.8.2"
pylint-django = "^2.4.4"
rope = "^0.11.0"
# formatting
black = {version = "^21.5b0", allow-prereleases = true}
isort = { version = "^5.5.4", extras = ["pyproject"] }
pre-commit = "^2.10.1"
# testing
Faker = "^8.1.2"
factory-boy = "^3.2.0"
pyhamcrest = "^1.9"
pytest = "^5.3.5"
pytest-django = "^3.8.0"
tblib = "^1.6.0"
# running
werkzeug = "^0.14.1"

[tool.isort]
# this configuration makes isort and black compatible
# https://black.readthedocs.io/en/stable/the_black_code_style.html#how-black-wraps-lines
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.pylint.master]
load-plugins = "pylint_django"

[tool.pylint."messages control"]
disable = "bad-continuation,too-many-ancestors,too-few-public-methods,duplicate-code"
enable = "useless-suppression"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
